refactor: divide regression to other package
